api/requirements.txt
fastapi
uvicorn
paho-mqtt
requests
sqlmodel
psycopg2-binary